Add files via upload
This is the code updated with functions so that one gets a clearer view of what is happening. Another thing worth mentioning is that this code has 25 weights of the NLMS and this code is in "NLMSthreeWeights" which is kinda misleading. The 25 weights was just a test to see if something would happen.